move survey into separate component
